+++ tests/TestCase/Registry/BroadcasterRegistryTest.php
@@ -28,7 +28,7 @@ class BroadcasterRegistryTest extends TestCase
      *
      * @return void
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -40,7 +40,7 @@ class BroadcasterRegistryTest extends TestCase
      *
      * @return void
      */
-    public function tearDown(): void
+    protected function tearDown(): void
     {
         parent::tearDown();